Fix connection of new position form to store
